refactor: move correct package structure QUZ-80
