+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/dto/GameCreateRequest.kt
@@ -1,10 +1,10 @@
 package com.grepp.quizy.game.api.game.dto
 
-import com.grepp.quizy.game.domain.GameLevel
-import com.grepp.quizy.game.domain.GameSubject
+import com.grepp.quizy.game.domain.game.GameLevel
+import com.grepp.quizy.game.domain.game.GameSubject
 
 class GameCreateRequest(
-        val subject: GameSubject,
-        val level: GameLevel,
-        val quizCount: Int,
+    val subject: GameSubject,
+    val level: GameLevel,
+    val quizCount: Int,
 ) {}
