+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/dto/MatchingRequest.kt
@@ -1,6 +1,6 @@
 package com.grepp.quizy.game.api.game.dto
 
-import com.grepp.quizy.game.domain.GameSubject
+import com.grepp.quizy.game.domain.game.GameSubject
 
 data class MatchingRequest(
     val userIds: List<Long>,
