--- game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/GameMessagePublisher.kt
@@ -1,5 +1,0 @@
-package com.grepp.quizy.game.domain
-
-interface GameMessagePublisher {
-    fun publish(message: GameMessage)
-}
