+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/Message.kt
@@ -1,5 +1,7 @@
 package com.grepp.quizy.game.domain
 
+import com.grepp.quizy.game.domain.game.*
+
 enum class MessageType {
     GAME_ROOM,
     CHAT,
