--- game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/QuizFetcher.kt
@@ -1,11 +1,0 @@
-package com.grepp.quizy.game.domain
-
-interface QuizFetcher {
-
-    fun fetchQuiz(
-        subject: GameSubject,
-        quizCount: Int,
-        level: GameLevel
-    ): GameQuiz
-
-}
