+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/GameApplicationEvent.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.game.domain
+package com.grepp.quizy.game.domain.game
 
 class GameStartEvent(
     val game: Game
