+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/GameMessagePublisher.kt
@@ -1,0 +1,7 @@
+package com.grepp.quizy.game.domain.game
+
+import com.grepp.quizy.game.domain.GameMessage
+
+interface GameMessagePublisher {
+    fun publish(message: GameMessage)
+}
