+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/GameReader.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.game.domain
+package com.grepp.quizy.game.domain.game
 
 import com.grepp.quizy.game.domain.exception.GameException
 import org.springframework.stereotype.Component
@@ -8,11 +8,11 @@ class GameReader(private val gameRepository: GameRepository) {
 
     fun read(id: Long): Game {
         return gameRepository.findById(id)
-                ?: throw GameException.GameNotFoundException
+            ?: throw GameException.GameNotFoundException
     }
 
     fun readByInviteCode(code: String): Game {
         return gameRepository.findByInviteCode(code)
-                ?: throw GameException.GameNotFoundException
+            ?: throw GameException.GameNotFoundException
     }
 }
