+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/GameSettingManager.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.game.domain
+package com.grepp.quizy.game.domain.game
 
 import org.springframework.stereotype.Component
 
@@ -6,27 +6,27 @@ import org.springframework.stereotype.Component
 class GameSettingManager(private val gameRepository: GameRepository) {
 
     fun updateSubject(
-            game: Game,
-            subject: GameSubject,
-            userId: Long,
+        game: Game,
+        subject: GameSubject,
+        userId: Long,
     ): Game {
         game.updateSubject(userId = userId, subject = subject)
         return gameRepository.save(game)
     }
 
     fun updateLevel(
-            game: Game,
-            level: GameLevel,
-            userId: Long,
+        game: Game,
+        level: GameLevel,
+        userId: Long,
     ): Game {
         game.updateLevel(userId = userId, level = level)
         return gameRepository.save(game)
     }
 
     fun updateQuizCount(
-            game: Game,
-            quizCount: Int,
-            userId: Long,
+        game: Game,
+        quizCount: Int,
+        userId: Long,
     ): Game {
         game.updateQuizCount(userId = userId, quizCount = quizCount)
         return gameRepository.save(game)
