+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/IdGenerator.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.game.domain
+package com.grepp.quizy.game.domain.game
 
 interface IdGenerator {
     fun generate(key: String): Long
