+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/quiz/QuizFetcher.kt
@@ -1,0 +1,14 @@
+package com.grepp.quizy.game.domain.quiz
+
+import com.grepp.quizy.game.domain.game.GameLevel
+import com.grepp.quizy.game.domain.game.GameSubject
+
+interface QuizFetcher {
+
+    fun fetchQuiz(
+        subject: GameSubject,
+        quizCount: Int,
+        level: GameLevel
+    ): GameQuiz
+
+}
