+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/FakeGameMessagePublisher.kt
@@ -1,5 +1,7 @@
 package com.grepp.quizy.game.domain
 
+import com.grepp.quizy.game.domain.game.GameMessagePublisher
+
 class FakeGameMessagePublisher() : GameMessagePublisher {
 
     private val messages = mutableListOf<GameMessage>()
