+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/FakeIdGenerator.kt
@@ -1,5 +1,7 @@
 package com.grepp.quizy.game.domain
 
+import com.grepp.quizy.game.domain.game.IdGenerator
+
 class FakeIdGenerator() : IdGenerator {
 
     private var currentId = 0L
