+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/GameMatchingServiceTest.kt
@@ -1,8 +1,9 @@
 package com.grepp.quizy.game.domain
 
-import com.grepp.quizy.game.domain.GameType.RANDOM
-import com.grepp.quizy.game.domain.PlayerRole.GUEST
-import com.grepp.quizy.game.domain.PlayerRole.HOST
+import com.grepp.quizy.game.domain.game.*
+import com.grepp.quizy.game.domain.game.GameType.RANDOM
+import com.grepp.quizy.game.domain.game.PlayerRole.GUEST
+import com.grepp.quizy.game.domain.game.PlayerRole.HOST
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
