+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/GamePlayerManagerTest.kt
@@ -1,6 +1,7 @@
 package com.grepp.quizy.game.domain
 
-import com.grepp.quizy.game.domain.GameType.PRIVATE
+import com.grepp.quizy.game.domain.game.*
+import com.grepp.quizy.game.domain.game.GameType.PRIVATE
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
