+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/GameReaderTest.kt
@@ -1,7 +1,8 @@
 package com.grepp.quizy.game.domain
 
-import com.grepp.quizy.game.domain.GameType.PRIVATE
 import com.grepp.quizy.game.domain.exception.GameException
+import com.grepp.quizy.game.domain.game.*
+import com.grepp.quizy.game.domain.game.GameType.PRIVATE
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
