+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/GameSettingTest.kt
@@ -1,5 +1,8 @@
 package com.grepp.quizy.game.domain
 
+import com.grepp.quizy.game.domain.game.GameLevel
+import com.grepp.quizy.game.domain.game.GameSetting
+import com.grepp.quizy.game.domain.game.GameSubject
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
