+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/GameTest.kt
@@ -1,9 +1,10 @@
 package com.grepp.quizy.game.domain
 
-import com.grepp.quizy.game.domain.GameType.PRIVATE
-import com.grepp.quizy.game.domain.PlayerRole.GUEST
-import com.grepp.quizy.game.domain.PlayerRole.HOST
 import com.grepp.quizy.game.domain.exception.GameException
+import com.grepp.quizy.game.domain.game.*
+import com.grepp.quizy.game.domain.game.GameType.PRIVATE
+import com.grepp.quizy.game.domain.game.PlayerRole.GUEST
+import com.grepp.quizy.game.domain.game.PlayerRole.HOST
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
