+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/InviteCodeTest.kt
@@ -1,5 +1,6 @@
 package com.grepp.quizy.game.domain
 
+import com.grepp.quizy.game.domain.game.InviteCode
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
