+++ game-service/game-domain/src/test/kotlin/com/grepp/quizy/game/domain/PlayerTest.kt
@@ -1,7 +1,9 @@
 package com.grepp.quizy.game.domain
 
-import com.grepp.quizy.game.domain.PlayerRole.GUEST
-import com.grepp.quizy.game.domain.PlayerRole.HOST
+import com.grepp.quizy.game.domain.game.Player
+import com.grepp.quizy.game.domain.game.PlayerRole.GUEST
+import com.grepp.quizy.game.domain.game.PlayerRole.HOST
+import com.grepp.quizy.game.domain.game.PlayerStatus
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
