+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/game/GameRedisIdGenerator.kt
@@ -1,12 +1,12 @@
 package com.grepp.quizy.game.infra.game
 
-import com.grepp.quizy.game.domain.IdGenerator
+import com.grepp.quizy.game.domain.game.IdGenerator
 import com.grepp.quizy.game.infra.redis.RedisINCRIdGenerator
 import org.springframework.stereotype.Component
 
 @Component
 class GameRedisIdGenerator(
-        private val redisINCRIdGenerator: RedisINCRIdGenerator
+    private val redisINCRIdGenerator: RedisINCRIdGenerator
 ) : IdGenerator {
 
     override fun generate(key: String): Long {
